+++ client/src/main/java/marksman/client/DataSource.java
@@ -1,0 +1,5 @@
+package marksman.client;
+
+public class DataSource {
+    public String userName;
+}
