+++ client/src/main/resources/marksman/client/login.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.HBox?>
 
 <BorderPane xmlns:fx="http://javafx.com/fxml/1" prefHeight="317.0" prefWidth="402.0"
-            xmlns="http://javafx.com/javafx/23.0.1" fx:controller="marksman.client.components.LoginComponent">
+            xmlns="http://javafx.com/javafx/23.0.1" fx:controller="marksman.client.login.Component">
     <center>
         <HBox alignment="CENTER" prefHeight="40.0" prefWidth="380.0" spacing="10.0">
             <children>
